Don't assume `APIResource.retrieve` returns type `cls`
Previously, we assumed that the inherited class of `APIResource` would
return its own type when calling `retrieve()`. This is no longer the case.

We have introduced the
[/v2/connections](https://developers.telnyx.com/docs/api/v2/connections/Connections#getConnection)
endpoint that can be thought of as a proxy endpoint to the various child
connection resources. This means that a call to

GET /v2/connections/{id}

Will NOT return resources with a `record_type` set to `connection` but
instead the `record_type` of the actual connection (`ip_connection`, for
example).